Update !delete confirmation to include deletion by assetid

+++ src/classes/Commands/main.ts
@@ -858,6 +858,16 @@ export = class Commands {
             const sku = ourInventory.findByAssetid(params.assetid);
 
             if (sku === null) {
+                if (params.i_am_sure !== 'yes_i_am') {
+                    this.bot.sendMessage(
+                        steamID,
+                        `/pre ⚠️ Are you sure that you want to delete the item with asset ID ${params.assetid}?` +
+                            `\n⚠️ This process is irreversible and will delete the item from your bot's backpack!` +
+                            `\n⚠️ If you are sure, try again with i_am_sure=yes_i_am as a parameter`
+                    );
+                    return;
+                }
+
                 this.bot.tf2gc.deleteItem(params.assetid, err => {
                     if (err) {
                         log.warn(`Error trying to delete ${params.assetid}: `, err);
